Create two Js files which is store common veriables and functions which is use in all other files
