+++ src/js/config.js
@@ -1,0 +1,2 @@
+export const API_URL ='https://forkify-api.herokuapp.com/api/v2/recipes';
+export const TIME_OUT = 10;
